requirements.txt
Django
djangorestframework
pymongo
djongo==1.3.7
djangorestframework-simplejwt
drf-yasg